brandme-agents/requirements.txt
# Copyright (c) Brand.Me, Inc. All rights reserved.

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25

# HTTP Client
httpx==0.26.0

# Logging & Monitoring
python-json-logger==2.0.7
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Utilities
python-dotenv==1.0.0
